moving admin repo to the correct folder
