+++ src/apps/admin-desktop/admin-desktop.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { UsersController } from './controllers/users.controller';
-import { UsersRepo } from 'src/data/repositories/admin-repos/users.repo';
+import { UsersRepo } from 'src/data/repositories/controllers-repos/admin-repos/users.repo';
 import { CraftmenController } from './controllers/craftmen.controller';
-import { CraftmenRepo } from 'src/data/repositories/admin-repos/craftmen.repo';
+import { CraftmenRepo } from 'src/data/repositories/controllers-repos/admin-repos/craftmen.repo';
 import { SettingsController } from './controllers/settings.controller';
-import { SettingsRepo } from 'src/data/repositories/admin-repos/settings.repo';
+import { SettingsRepo } from 'src/data/repositories/controllers-repos/admin-repos/settings.repo';
 import { DashboardController } from './controllers/dashboard.controller';
-import { DashboardRepo } from 'src/data/repositories/admin-repos/dashboard.repo';
+import { DashboardRepo } from 'src/data/repositories/controllers-repos/admin-repos/dashboard.repo';
 import { AnalyzesController } from './controllers/analyzes.controller';
 
 @Module({
